requirements.txt
Django~=5.2.4
psycopg2-binary~=2.9.10
pandas~=2.2.0 # Loosening from 2.3.0 as it's very new, 2.2 is safer for wider compatibility
openpyxl~=3.1.2
PyPDF2~=3.0.1
pdfplumber~=0.11.0
gunicorn>=21.2.0 # Added gunicorn
python-dotenv~=1.0.0 # For managing environment variables locally, useful with Docker
pillow~=10.2.0 # pdfplumber dependency, good to make explicit
numpy~=1.26.4 # pandas dependency
reportlab~=4.0.8 # For PDF generation

# Google Cloud dependencies
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.20.0
django-storages[google]>=1.14.2

# Database URL parsing for Cloud SQL
dj-database-url>=2.1.0

# WhiteNoise for static file fallback
whitenoise>=6.7.0

# Redis cache (optional but recommended for production)
django-redis>=5.4.0
redis>=5.0.0

# CORS handling for API endpoints
django-cors-headers>=4.3.0

# Security and monitoring
sentry-sdk[django]>=1.40.0

# Email backend for notifications (optional)
django-anymail[mailgun]>=10.3.0

# Health check dependencies
psutil>=5.9.0
